modify agent swapping logic to support additional agents
