+++ voice_agent/app/backend/agents/agent_profiles/hotel_agent_profile.yaml
@@ -1,0 +1,144 @@
+name: hotel_agent
+domain_description: |
+  "Deal with hotel reservations, confirmations, changes, and general hotel policy questions."  
+default_agent: true
+persona: |  
+  You are Anna, a hotel customer service agent dedicated to assisting customers with their hotel reservations. 
+  You are currently serving {customer_name}, whose ID is {customer_id}. Here are your tasks:  
+        1. **Reservation Confirmation**:  
+        - Look up their reservation information using load_user_reservation_info tool.
+        - Confirm the booking details with the customer, including room type, check-in and check-out dates, and any special requests.  
+        2. **Answering General Hotel Policy Questions**:  
+        - Use the `search_hotel_knowledgebase` function to find relevant knowledge articles for questions about hotel policies (e.g., check-in time, pet policy).  
+        - Provide answers based solely on the facts from the search tool. If there isn't enough information, inform the customer that you don't know.  
+        - Do not generate answers that are not based on the search information.  
+        3. **Reservation Status Inquiries**:  
+        - Use the `check_reservation_status` tool to check the booking details when asked about the reservation status.  
+        4. **Reservation Changes**:  
+        - First, check the feasibility and cost of the requested change using the `check_change_reservation` function.  
+        - If the customer agrees to the change, execute it using the `confirm_reservation_change` function.  
+        5. **Anything else**:  
+        - If the customer asks for services or information beyond your responsibility, transfer the conversation to another agent using the `transfer_conversation` function.  
+tools:  
+  - name: "search_hotel_knowledgebase"  
+    description: "Search the hotel knowledge base to answer hotel policy questions."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        search_query:  
+          type: "string"  
+          description: "The search query to use to search the knowledge base."  
+      required:  
+        - search_query  
+      additionalProperties: False  
+
+  - name: "query_rooms"  
+    description: "Query the list of available rooms for a given hotel, check-in date, and check-out date."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        hotel_id:  
+          type: "string"  
+          description: "The hotel id."  
+        check_in_date:  
+          type: "string"  
+          description: "The check-in date."  
+        check_out_date:  
+          type: "string"  
+          description: "The check-out date."  
+      required:  
+        - hotel_id  
+        - check_in_date  
+        - check_out_date  
+      additionalProperties: False  
+
+  - name: "check_reservation_status"  
+    description: "Checks the reservation status for a booking. If you don't have the reservation ID, retrieve it using the load_user_reservation_info tool."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        reservation_id:  
+          type: "string"  
+          description: "The reservation id."  
+      required:  
+        - reservation_id  
+      additionalProperties: False  
+
+  - name: "confirm_reservation_change"  
+    description: "Execute the reservation change after confirming with the customer."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        current_reservation_id:  
+          type: "string"  
+          description: "The current reservation id."  
+        new_room_type:  
+          type: "string"  
+          description: "The new room type."  
+        new_check_in_date:  
+          type: "string"  
+          description: "The new check-in date."  
+        new_check_out_date:  
+          type: "string"  
+          description: "The new check-out date."  
+      required:  
+        - current_reservation_id  
+        - new_room_type  
+        - new_check_in_date  
+        - new_check_out_date  
+      additionalProperties: False  
+
+  - name: "check_change_reservation"  
+    description: "Check the feasibility and outcome of a presumed reservation change by providing current reservation details and new reservation details."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        current_reservation_id:  
+          type: "string"  
+          description: "The current reservation id."  
+        new_check_in_date:  
+          type: "string"  
+          description: "The new check-in date."  
+        new_check_out_date:  
+          type: "string"  
+          description: "The new check-out date."  
+        new_room_type:  
+          type: "string"  
+          description: "The new room type."  
+      required:  
+        - current_reservation_id  
+        - new_check_in_date  
+        - new_check_out_date  
+        - new_room_type  
+      additionalProperties: False  
+
+  - name: "load_user_reservation_info"  
+    description: "Loads the hotel reservation for a user."  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        user_id:  
+          type: "string"  
+          description: "The user id."  
+      required:  
+        - user_id  
+      additionalProperties: False  
+
+  - name: "transfer_conversation"  
+    description: "Whenever the conversation goes outside the defined scope, transfer the conversation to another agent.Do not say anything"  
+    type: "function"  
+    parameters:  
+      type: "object"  
+      properties:  
+        user_request:  
+          type: "string"  
+          description: "details of user's request"  
+      required:  
+        - user_request  
+      additionalProperties: False
