+++ voice_agent/app/backend/agents/agent_profiles/incomplete/car_rental_agent_profile.yaml
@@ -1,0 +1,4 @@
+name: car_rental_agent
+domain_description: |
+  "Deal with car rental reservations, confirmations, changes, and general car rental policy questions."
+default_agent: false
